Put strategy constants, types, validators into a single file
