--- frontend/src/metabase/admin/performance/constants.tsx
@@ -1,1 +1,0 @@
-export const rootId = 0;
