--- frontend/src/metabase/admin/performance/utils.ts
@@ -1,25 +1,0 @@
-import type { SchemaObjectDescription } from "yup/lib/schema";
-
-import type { StrategyType, type Strategy } from "./types";
-import { Strategies } from "./types";
-
-export const getStrategyLabel = (strategy?: Strategy) => {
-  return strategy ? Strategies[strategy.type].label : null;
-};
-
-export const getShortStrategyLabel = (strategy?: Strategy) => {
-  if (!strategy) {
-    return null;
-  }
-  const type = Strategies[strategy.type];
-  return type.shortLabel ?? type.label;
-};
-
-export const getFieldsForStrategyType = (strategyType: StrategyType) => {
-  const strategy = Strategies[strategyType];
-  const validationSchemaDescription =
-    strategy.validateWith.describe() as SchemaObjectDescription;
-  const fieldRecord = validationSchemaDescription.fields;
-  const fields = Object.keys(fieldRecord);
-  return fields;
-};
